igfx: Unconditionally enforce complete modesetting on Skylake and newer
The framebuffer controller can perform panel fitter, partial, or a
complete modeset (see AppleIntelFramebufferController::hwSetMode).
In a dual-monitor CFL DVI+HDMI setup, only HDMI output was working after
boot: it was observed that for HDMI framebuffer a complete modeset
eventually occured, but for DVI it never did until after sleep and wake
sequence.

Function AppleIntelFramebufferController::hwRegsNeedUpdate checks
whether a complete modeset needs to be issued. It does so by comparing
sets of pipes and transcoder parameters. For some reason, the result was
never true in the above scenario, so a complete modeset never occured.
Consequently, AppleIntelFramebufferController::LightUpTMDS was never
called for that framebuffer.

Patching hwRegsNeedUpdate to always return true seems to be a rather
safe solution to that. Note that the root cause of the problem is
somewhere deeper.