Add codec hardware info spoofing for AMD VCE
When an AMD card is PID spoofed, video encoding does not work because in
AMDRadeonVADriver, `VAAcceleratorInfo::identify` gets the PID by calling
up to AMDRadeonX4000.kext (in a method
`AMDRadeonX4000_AMDAccelVideoContext::getHWInfo` which gets the PID from
`AMDRadeonX4000_AMDHardware::getHWInfo` which is set up by
`AMDRadeonX4000_AMDHardware::initHWInfo` which copies the value from
`AMDRadeonX4000_AMDHardware::init` which reads the PCI registers).

We patch `getHWInfo` to return the spoofed PID (from `device-id`) and this
way, when AppleGVA queries AMDRadeonVADriver, it will no longer return an
error.

Additionally, Shiki patches can be applied to get DRM to work.